refactor(cron): type transaction hooks on the internal store writer
