+++ config/assertion-safety-baseline.txt
@@ -2722,7 +2722,6 @@ src/cron/store.ts	1
 src/cron/store/delivery-codec.ts	2
 src/cron/store/quarantine.ts	1
 src/cron/store/row-codec.ts	4
-src/cron/store/transaction-hooks.ts	1
 src/cron/task-run-detail.ts	2
 src/daemon/arg-split.ts	2
 src/daemon/launchd-install.ts	7
