+++ src/cron/store.test.ts
@@ -25,7 +25,6 @@ import {
   saveCronQuarantinedJobs,
   saveCronStore,
 } from "./store.js";
-import { saveCronJobsStoreWithTransactionHooks } from "./store/transaction-hooks.js";
 import type { CronStoreFile } from "./types.js";
 
 let fixtureRoot = "";
@@ -487,7 +486,7 @@ describe("cron store", () => {
     const { storePath } = await makeStorePath();
     const store = makeStore("post-commit-hook", true);
     const afterCommit = vi.fn();
-    await saveCronJobsStoreWithTransactionHooks(storePath, store, undefined, { afterCommit });
+    await saveCronJobsStore(storePath, store, { transactionHooks: { afterCommit } });
     expect(afterCommit).toHaveBeenCalledOnce();
 
     const database = openOpenClawStateDatabase().db;
@@ -496,7 +495,7 @@ describe("cron store", () => {
     );
     try {
       await expect(
-        saveCronJobsStoreWithTransactionHooks(storePath, store, undefined, { afterCommit }),
+        saveCronJobsStore(storePath, store, { transactionHooks: { afterCommit } }),
       ).rejects.toThrow("cron update rejected");
       expect(afterCommit).toHaveBeenCalledOnce();
     } finally {
@@ -1795,11 +1794,14 @@ describe("cron jobs fingerprint guard", () => {
     await saveCronStore(storePath, { version: 1, jobs: [jobB, jobA] });
 
     await expect(
-      saveCronJobsStoreWithTransactionHooks(
+      saveCronJobsStore(
         storePath,
         { version: 1, jobs: [jobA, jobB] },
-        undefined,
-        { beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, staleFingerprint) },
+        {
+          transactionHooks: {
+            beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, staleFingerprint),
+          },
+        },
       ),
     ).rejects.toBeInstanceOf(CronJobsStoreChangedError);
     expect((await loadCronStore(storePath)).jobs.map((job) => job.id)).toEqual(["job-b", "job-a"]);
@@ -1809,11 +1811,14 @@ describe("cron jobs fingerprint guard", () => {
       "fingerprint after concurrent reorder",
     );
     expect(freshFingerprint).not.toBe(staleFingerprint);
-    await saveCronJobsStoreWithTransactionHooks(
+    await saveCronJobsStore(
       storePath,
       { version: 1, jobs: [jobA, jobB] },
-      undefined,
-      { beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, freshFingerprint) },
+      {
+        transactionHooks: {
+          beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, freshFingerprint),
+        },
+      },
     );
     expect((await loadCronStore(storePath)).jobs.map((job) => job.id)).toEqual(["job-a", "job-b"]);
   });
@@ -1848,12 +1853,12 @@ describe("cron jobs fingerprint guard", () => {
     expect((await loadCronJobsStoreWithConfigJobs(storePath)).jobsFingerprint).toBe(fingerprint);
     const repair = structuredClone(store);
     expectDefined(repair.jobs[0], "repair job").enabled = false;
-    await saveCronJobsStoreWithTransactionHooks(
-      storePath,
-      repair,
-      { preserveRuntimeState: true },
-      { beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, fingerprint) },
-    );
+    await saveCronJobsStore(storePath, repair, {
+      preserveRuntimeState: true,
+      transactionHooks: {
+        beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, fingerprint),
+      },
+    });
 
     const repaired = expectDefined((await loadCronStore(storePath)).jobs[0], "repaired job");
     expect(repaired.enabled).toBe(false);
@@ -1891,12 +1896,12 @@ describe("cron jobs fingerprint guard", () => {
       toolsAllow: ["read"],
     };
 
-    await saveCronJobsStoreWithTransactionHooks(
-      storePath,
-      repair,
-      { preserveRuntimeState: true },
-      { beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, fingerprint) },
-    );
+    await saveCronJobsStore(storePath, repair, {
+      preserveRuntimeState: true,
+      transactionHooks: {
+        beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, fingerprint),
+      },
+    });
 
     const repaired = expectDefined((await loadCronStore(storePath)).jobs[0], "repaired job");
     expect(repaired.runtimeAuthority).toBeUndefined();
@@ -1919,12 +1924,12 @@ describe("cron jobs fingerprint guard", () => {
     const repair = structuredClone(store);
     expectDefined(repair.jobs[0], "repair job").enabled = false;
 
-    await saveCronJobsStoreWithTransactionHooks(
-      storePath,
-      repair,
-      { preserveRuntimeState: true },
-      { beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, fingerprint) },
-    );
+    await saveCronJobsStore(storePath, repair, {
+      preserveRuntimeState: true,
+      transactionHooks: {
+        beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, fingerprint),
+      },
+    });
 
     const repaired = expectDefined((await loadCronStore(storePath)).jobs[0], "repaired job");
     expect(repaired.enabled).toBe(false);
@@ -1950,12 +1955,12 @@ describe("cron jobs fingerprint guard", () => {
     const loaded = await loadCronJobsStoreWithConfigJobs(storePath);
     const fingerprint = expectDefined(loaded.jobsFingerprint, "legacy authority fingerprint");
 
-    await saveCronJobsStoreWithTransactionHooks(
-      storePath,
-      loaded.store,
-      { preserveRuntimeState: true },
-      { beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, fingerprint) },
-    );
+    await saveCronJobsStore(storePath, loaded.store, {
+      preserveRuntimeState: true,
+      transactionHooks: {
+        beforeWrite: (db) => assertCronJobsStoreUnchanged(db, storePath, fingerprint),
+      },
+    });
 
     expect((await loadCronStore(storePath)).jobs[0]?.runtimeAuthority).toEqual(
       job.runtimeAuthority,
