+++ src/cron/store.ts
@@ -266,6 +266,7 @@ type SaveCronStoreOptions = {
 };
 
 type SaveCronJobsStoreOptions = SaveCronStoreOptions & {
+  transactionHooks?: CronStoreTransactionHooks;
   quarantine?: {
     entries: readonly (QuarantinedCronConfigJob | CronQuarantinedJob)[];
     nowMs: number;
@@ -278,10 +279,6 @@ type CronStoreReplacementOptions = Pick<
   SaveCronJobsStoreOptions,
   "deleteQuarantineEntries" | "preserveRuntimeState" | "quarantine"
 >;
-
-type SaveCronJobsStoreInternalOptions = SaveCronJobsStoreOptions & {
-  transactionHooks?: CronStoreTransactionHooks;
-};
 
 function mergeCronRuntimeChanges(
   previous: CronJobState,
@@ -441,11 +438,6 @@ export async function saveCronJobsStore(
   storePath: string,
   store: CronStoreFile,
   opts?: SaveCronJobsStoreOptions,
-): Promise<void>;
-export async function saveCronJobsStore(
-  storePath: string,
-  store: CronStoreFile,
-  opts?: SaveCronJobsStoreInternalOptions,
 ): Promise<void> {
   const resolvedStorePath = path.resolve(storePath);
   const storeKey = cronStoreKey(resolvedStorePath);
